+++ pkg/integer/functions_test.go
@@ -12,7 +12,7 @@ import (
 func TestQuantization_Quantize(t *testing.T) {
 	q := NewQuantization(7, 50)
 	a := q.Quantize(3.5678)
-	assert.Equal(t, a.q, 9)
+	assert.Equal(t, a.q, int32(9))
 	x := q.Dequantize(a.q)
 	assert.InDelta(t, x, 3.5678, 1.0e-1)
 }
@@ -20,13 +20,13 @@ func TestQuantization_Quantize(t *testing.T) {
 func TestQuantiztion_IntegerGelu(t *testing.T) {
 	q := NewQuantization(7, 50)
 	a := q.Quantize(0.55)
-	assert.Equal(t, a.q, 1)
+	assert.Equal(t, a.q, int32(1))
 	gelu := q.IntegerGelu(a.q)
-	assert.Equal(t, gelu.q, -54)
+	assert.Equal(t, gelu.q, int32(-54))
 	assert.Equal(t, gelu.scaling, float32(-0.0044071344))
 	b := q.Quantize(-1)
 	gelu = q.IntegerGelu(b.q)
-	assert.Equal(t, gelu.q, 48)
+	assert.Equal(t, gelu.q, int32(48))
 	assert.Equal(t, gelu.scaling, float32(-0.0044071344))
 }
 
@@ -47,7 +47,7 @@ func TestQuantiztion_IntegerExp(t *testing.T) {
 
 func TestQuantiztion_IntegerSoftmax(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v := []int{-45, 98, -491}
+	v := []int32{-45, 98, -491}
 	softmax := q.IntSoftmax(v)
 	s := float32(softmax[0].q) * softmax[0].scaling
 	assert.InDelta(t, s, float32(0.1492918), 1.0e-6)
@@ -59,20 +59,20 @@ func TestQuantiztion_IntegerSoftmax(t *testing.T) {
 
 func TestQuantiztion_IntegerSquareRoot(t *testing.T) {
 	sqrt := IntSqrt(40)
-	assert.Equal(t, sqrt, 6)
+	assert.Equal(t, sqrt, int32(6))
 	sqrt = IntSqrt(0)
-	assert.Equal(t, sqrt, 0)
+	assert.Equal(t, sqrt, int32(0))
 	sqrt = IntSqrt(9)
-	assert.Equal(t, sqrt, 3)
+	assert.Equal(t, sqrt, int32(3))
 	sqrt = IntSqrt(100)
-	assert.Equal(t, sqrt, 10)
+	assert.Equal(t, sqrt, int32(10))
 	sqrt = IntSqrt(90)
-	assert.Equal(t, sqrt, 9)
+	assert.Equal(t, sqrt, int32(9))
 }
 
 func TestQuantization_IntegerLayerNorm(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v := []int{-45, 98, -491}
+	v := []int32{-45, 98, -491}
 	norm := q.IntNormalization(v)
 	s := float32(norm[0].q) * norm[0].scaling
 	assert.InDelta(t, s, float32(0.40293040), 1.0e-6)
@@ -84,68 +84,68 @@ func TestQuantization_IntegerLayerNorm(t *testing.T) {
 
 func TestQuantization_Transpose(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	c := Transpose(a)
-	assert.Equal(t, c.matrix[0], []int{2, 3})
-	assert.Equal(t, c.matrix[1], []int{-2, 4})
-	assert.Equal(t, c.matrix[2], []int{4, -3})
+	assert.Equal(t, c.matrix[0], []int32{2, 3})
+	assert.Equal(t, c.matrix[1], []int32{-2, 4})
+	assert.Equal(t, c.matrix[2], []int32{4, -3})
 	assert.InDelta(t, c.scaling, float32(0.012210012), 1.0e-6)
 }
 
 func TestQuantization_LinearMul(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
-	v2 := []int{2, 4, 8, 7, -1, 0}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
+	v2 := []int32{2, 4, 8, 7, -1, 0}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	b := q.GetQuantizedMatrixFromInt(3, 2, v2)
 	c := Mul(a, b)
-	assert.Equal(t, c.matrix[0], []int{-16, -6})
-	assert.Equal(t, c.matrix[1], []int{41, 40})
+	assert.Equal(t, c.matrix[0], []int32{-16, -6})
+	assert.Equal(t, c.matrix[1], []int32{41, 40})
 	assert.InDelta(t, c.scaling, float32(0.00014908), 1.0e-6)
 }
 
 func TestQuantization_LinearAdd(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
-	v2 := []int{2, 4, 8, 7, -1, 0}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
+	v2 := []int32{2, 4, 8, 7, -1, 0}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	b := q.GetQuantizedMatrixFromInt(2, 3, v2)
 	c := Add(a, b)
-	assert.Equal(t, c.matrix[0], []int{4, 2, 12})
-	assert.Equal(t, c.matrix[1], []int{10, 3, -3})
+	assert.Equal(t, c.matrix[0], []int32{4, 2, 12})
+	assert.Equal(t, c.matrix[1], []int32{10, 3, -3})
 	assert.InDelta(t, c.scaling, float32(0.012210012), 1.0e-6)
 }
 
 func TestQuantization_LinearProd(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
-	v2 := []int{2, 4, 8, 7, -1, 0}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
+	v2 := []int32{2, 4, 8, 7, -1, 0}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	b := q.GetQuantizedMatrixFromInt(2, 3, v2)
 	c := Prod(a, b)
-	assert.Equal(t, c.matrix[0], []int{4, -8, 32})
-	assert.Equal(t, c.matrix[1], []int{21, -4, 0})
+	assert.Equal(t, c.matrix[0], []int32{4, -8, 32})
+	assert.Equal(t, c.matrix[1], []int32{21, -4, 0})
 	assert.InDelta(t, c.scaling, float32(0.00014908), 1.0e-6)
 }
 
 func TestQuantization_LinearProdScalar(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
 	s := QuantizedInt{
 		q:       2,
 		scaling: q.scaling,
 	}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	c := ProdScalar(a, s)
-	assert.Equal(t, c.matrix[0], []int{4, -4, 8})
-	assert.Equal(t, c.matrix[1], []int{6, 8, -6})
+	assert.Equal(t, c.matrix[0], []int32{4, -4, 8})
+	assert.Equal(t, c.matrix[1], []int32{6, 8, -6})
 	assert.InDelta(t, c.scaling, float32(0.00014908), 1.0e-6)
 }
 
 func TestQuantization_LinearDequantize(t *testing.T) {
 	q := NewQuantization(12, 50)
-	v1 := []int{2, -2, 4, 3, 4, -3}
+	v1 := []int32{2, -2, 4, 3, 4, -3}
 	a := q.GetQuantizedMatrixFromInt(2, 3, v1)
 	f := q.DequantizeMatrix(a)
 	assert.InDeltaSlice(t, f[0], []float32{0.024420024, -0.024420024, 0.048840048}, 1.0e-6)
